Replaced Expression Language with Andrzej's
It is possible to write every expression you want and there is a
temporal auto-completion for the states as well. Watch out the
expressions inside ifs; from now on they require a comma (",") at their
end in order to distinguish them. It is necessary for Xtext but also
good looking and helpful for the user.